% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/real.R
\name{COD}
\alias{COD}
\title{Coefficient of Dispersion}
\usage{
COD(x, y)
}
\arguments{
\item{x}{Appraisal value}

\item{y}{Sales price}
}
\value{
COD
}
\description{
Measurement of horizontal equity. Where horizontal equity is the deviation
of like properties from one another.
}
\examples{
x <- c(25500,57000,39000,90000,51000,93000,49500);
y <- c(75000,150000,90000,180000,90000,150000,75000);
codvalue <- COD(x,y);
}
