% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/real.R
\name{Constant.In.Data}
\alias{Constant.In.Data}
\title{Constant.In.Data}
\usage{
Constant.In.Data(dataf)
}
\arguments{
\item{dataf}{data.framed}
}
\value{
Returns columns with only 1 unique value
}
\description{
Finds columns in a data.frame that have only 1 unique value
}
