% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/real.R
\name{Data.Frame.Missing.Data}
\alias{Data.Frame.Missing.Data}
\title{Data.Frame.Missing.Data}
\usage{
Data.Frame.Missing.Data(dataf, Percent.Missing = -1)
}
\arguments{
\item{dataf}{data.frame}

\item{Percent_Missing}{An optional argument that will set a lower
bound of what percentage of data needs to be missing in order for it to be returned}
}
\value{
Returns a data.frame containing missing value information
}
\description{
This function accepts a data.frame and will return a data.frame
with all the columns and the percent of missing data that can be found in that column
}
