% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/real.R
\name{ID.In.Data}
\alias{ID.In.Data}
\title{ID.In.Data}
\usage{
ID.In.Data(dataf)
}
\arguments{
\item{dataf}{data.framed}
}
\value{
Returns possible ID columns
}
\description{
Finds columns in a data.frame that have no non-unique values (possible ID's)
}
