% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/real.R
\name{poly.trend}
\alias{poly.trend}
\title{Polynomial Time Trend Fit}
\usage{
poly.trend(data1)
}
\arguments{
\item{x}{Appraisal value}

\item{y}{Sales price}
}
\value{
rate table and a plot of the time trend fit
}
\description{
Calculates a polynomial fit based on the regression results of the relationship
of the sales price to time. Input must be a data.frame with the first column being sales
and the second being time i,e, months. Time must have a origin at 0.
}
\examples{
x <- c(10000,12000,11000,44000,35000,38000,89000,94000,102000,120000,125000,132000
,135000,137000,142000,145000,147000,149000);
y <- c(5,5,5,4,4,4,3,3,3,2,2,2,1,1,1,0,0,0);
df <- data.frame(salesprice = x,months = y)
poly_fit <- poly.trend(df);
}
