% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/real.R
\name{PRB}
\alias{PRB}
\title{Price Related Bias}
\usage{
PRB(x, y)
}
\arguments{
\item{x}{Appraisal value}

\item{y}{Sales price}
}
\value{
PRB
}
\description{
Measurement of the linear relationship of the appraisal value to the sales price.
Developed by Bob.
}
\examples{
x <- c(25500,57000,39000,90000,51000,93000,49500);
y <- c(75000,150000,90000,180000,90000,150000,75000);
prbvalue <- PRB(x,y);
}
