% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/real.R
\name{PRD}
\alias{PRD}
\title{Price Related Differential}
\usage{
PRD(x, y)
}
\arguments{
\item{x}{Appraisal value}

\item{y}{Sales price}
}
\value{
PRD
}
\description{
Measurement of vertical equity. Where vertical equity is the mean divided mby the
weighted mean, which gives equity performance of appraisal to sales ratios of lower
to higher valued properties.
}
\examples{
x <- c(25500,57000,39000,90000,51000,93000,49500);
y <- c(75000,150000,90000,180000,90000,150000,75000);
prdvalue <- PRD(x,y);
}
