% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/real.R
\name{REALsummary}
\alias{REALsummary}
\title{REAL Function Summary}
\usage{
REALsummary(a, s, ul = 1.7, ll = 0.3, year, df = FALSE)
}
\arguments{
\item{x}{Appraisal value}

\item{y}{Sales price}
}
\value{
summary results
}
\description{
returns a console output of all the appraisal calculations
}
\examples{
x <- c(25500,57000,39000,90000,51000,93000,49500);
y <- c(75000,150000,90000,180000,90000,150000,75000);
real_sum <- REALsummary(x,y);
}
