% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/real.R
\name{Weighted_Mean}
\alias{Weighted_Mean}
\title{Weighted Mean}
\usage{
Weighted_Mean(x, y)
}
\arguments{
\item{x}{Appraisal value}

\item{y}{Sales price}
}
\value{
weighted mean
}
\description{
Returns the weighted mean
}
\examples{
x <- c(25500,57000,39000,90000,51000,93000,49500);
y <- c(75000,150000,90000,180000,90000,150000,75000);
wmvalue <- Weighted_Mean(x,y);
}
