% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/real.R
\name{Weighted_Median}
\alias{Weighted_Median}
\title{Weighted Median}
\usage{
Weighted_Median(x, y)
}
\arguments{
\item{x}{Appraisal value}

\item{y}{Sales price}
}
\value{
weighted median
}
\description{
Returns the weighted median, used by certain bodies for appraisal. The
mathematical utility of such an equation is suspect.
}
\examples{
x <- c(25500,57000,39000,90000,51000,93000,49500);
y <- c(75000,150000,90000,180000,90000,150000,75000);
wmedianvalue <- Weighted_Median(x,y);
}
