% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/real.R
\name{REALsummary}
\alias{REALsummary}
\title{REAL Function Summary}
\usage{
REALsummary(a, s, trim = "YES", ul = 1.7, ll = 0.3, year, df = FALSE)
}
\arguments{
\item{trim}{YES is default. Enables ratio trimming}

\item{ul}{Upper limit for ratios. 1.7 is the default value}

\item{ll}{Lower limit for ratios. .3 is the default value}

\item{year}{If years are included, runs summary analysis broken up by years}

\item{df}{Default is FALSE. Run data.frame}

\item{x}{Appraisal value}

\item{y}{Sales price}
}
\value{
summary results
}
\description{
returns a console output of all the appraisal calculations
}
\examples{
x <- c(25500,57000,39000,90000,51000,93000,49500);
y <- c(75000,150000,90000,180000,90000,150000,75000);
real_sum <- REALsummary(x,y);
}
